refactor: #41 textVariants.ts -> variants.ts로 파일명 변경
