+++ ds/components/atoms/text/Text.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import "../../../styles/globals.css"
-import { textVariants } from "../../../styles/tokens/text/textVariants"
+import { textVariants } from "../../../styles/tokens/text/variants"
 import { TextProps } from "./Text.types"
 
 const Text: React.FC<TextProps> = ({
